ASF-RAG-backend/RAG_M/src/vectorstore/vector_store.py
import os
import warnings
from functools import lru_cache
from typing import List, Optional
import tempfile
import uuid
import shutil

from dotenv import load_dotenv
from langchain.docstore.document import Document
from langchain_huggingface import HuggingFaceEmbeddings
from langchain_community.vectorstores import FAISS

# 修改导入语句，使用正确的绝对导入
import sys
sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
from models.model_config import get_model_config

import os
import json


os.environ["HF_ENDPOINT"] = "https://hf-mirror.com"

load_dotenv()

#VECTORSTORE_PATH = os.getenv("VECTORSTORE_PATH")

class VectorStoreManager:
    """Manager for creating and loading FAISS vector stores"""
    def __init__(self, docs_dir: str = None):
        """Initialize vector store manager with embedding model from config file"""
        self._embeddings: Optional[HuggingFaceEmbeddings] = None
        # 优先使用配置文件中的模型，否则使用统一配置
        self._embedding_model = self._load_embedding_config(docs_dir)
        if not self._embedding_model:
            model_config = get_model_config()
            self._embedding_model = model_config.embedding_model

    def _load_embedding_config(self, docs_dir: str) -> str:
        """从knowledge_data.json加载embedding模型配置"""
        default_embedding = "BAAI/bge-small-zh-v1.5"
        if not docs_dir:
            print(f"使用默认的 embedding 模型: {default_embedding}")
            return default_embedding
        
        # 获取knowledge_data.json路径
        config_path = os.path.join(docs_dir, "knowledge_data.json")
        
        try:
            if os.path.exists(config_path):
                with open(config_path, 'r', encoding='utf-8') as f:
                    config = json.load(f)
                    print(f"加载embedding配置成功，使用模型: {config.get('embedding_model', default_embedding)}")
                    return config.get("embedding_model", default_embedding)
            return default_embedding
        except Exception as e:
            print(f"加载embedding配置失败: {e}")
            return default_embedding


    @property
    @lru_cache(maxsize=1)
    def embeddings(self) -> HuggingFaceEmbeddings:
        """Lazy load and cache embeddings model"""
        if not self._embeddings:
            self._embeddings = HuggingFaceEmbeddings(
                model_name=self._embedding_model,
                encode_kwargs={"normalize_embeddings": True}
            )
        return self._embeddings

    def _get_ascii_safe_temp_dir(self) -> str:
        """
        Return an ASCII-only temporary directory for FAISS write operations.
        FAISS on Windows can fail when writing directly to non-ASCII paths.
        """
        candidates = []
        env_temp = os.environ.get("FAISS_TEMP_DIR")
        if env_temp:
            candidates.append(env_temp)
        candidates.extend(["C:/faiss_temp", "D:/faiss_temp"])

        for base_dir in candidates:
            normalized = os.path.abspath(os.path.normpath(base_dir))
            if not normalized.isascii():
                continue
            try:
                os.makedirs(normalized, exist_ok=True)
                temp_dir = os.path.join(normalized, f"faiss_save_{uuid.uuid4().hex}")
                os.makedirs(temp_dir, exist_ok=True)
                return temp_dir
            except OSError:
                continue

        fallback = tempfile.mkdtemp(prefix="faiss_save_")
        if fallback.isascii():
            return fallback
        raise RuntimeError(
            "Unable to create an ASCII-only temporary directory for FAISS. "
            "Please set FAISS_TEMP_DIR to an ASCII path (e.g. C:/faiss_temp)."
        )

    def create_vectorstore(self, documents: List[Document], save_path: str) -> FAISS:
        """Create and save a FAISS vector store from documents"""
        if not documents:
            raise ValueError("No documents provided to create vector store")
        
        save_path = os.path.abspath(os.path.normpath(save_path))
        print(f"Attempting to create vector store at: {save_path}")
        
        try:
            # 创建向量存储
            print(f"Creating FAISS vector store with {len(documents)} documents...")
            vectorstore = FAISS.from_documents(documents, self.embeddings)
            
            # 保存向量存储前确保目录存在
            print(f"Ensuring save directory exists: {save_path}")
            os.makedirs(save_path, exist_ok=True)
            
            # 创建ASCII安全临时目录（避免Windows下FAISS写入中文路径失败）
            temp_dir = self._get_ascii_safe_temp_dir()
            
            # 验证临时目录是否可写
            test_file = os.path.join(temp_dir, '.write_test')
            try:
                with open(test_file, 'w') as f:
                    f.write('test')
                os.remove(test_file)
                print("Temporary directory write permissions verified")
            except Exception as e:
                raise RuntimeError(f"Cannot write to temporary directory: {str(e)}")
            
            # 保存到临时目录
            print(f"Saving vector store to temporary directory: {temp_dir}")
            vectorstore.save_local(temp_dir)
            
            # 移动文件到最终位置
            for file in os.listdir(temp_dir):
                src = os.path.join(temp_dir, file)
                dst = os.path.join(save_path, file)
                print(f"Moving file: {src} to {dst}")
                shutil.move(src, dst)
            
            # 清理临时目录
            shutil.rmtree(temp_dir)
            
            print(f"Vector store successfully created and saved to {save_path}")
            
            # 验证文件是否创建成功
            required_files = ['index.faiss', 'index.pkl']
            for file in required_files:
                file_path = os.path.join(save_path, file)
                if not os.path.exists(file_path):
                    raise RuntimeError(f"Expected file not found after save: {file_path}")
                print(f"Verified file exists: {file_path}")
            
            return vectorstore
            
        except Exception as e:
            print(f"Error creating vector store: {str(e)}")
            print(f"Current working directory: {os.getcwd()}")
            print(f"Save path exists: {os.path.exists(save_path)}")
            if os.path.exists(save_path):
                print(f"Save path is writable: {os.access(save_path, os.W_OK)}")
                print(f"Contents of save directory: {os.listdir(save_path)}")
            raise

    
    def initialize_vectorstore(self, save_path: str):
        """Initialize an empty vector store with required files"""
        # 规范化路径并转换为绝对路径
        save_path = os.path.abspath(os.path.normpath(save_path))
        
        # 确保目录存在且有写权限
        try:
            # 创建目录（如果不存在）
            os.makedirs(save_path, exist_ok=True)
            
            # 再次检查目录是否存在
            if not os.path.exists(save_path):
                raise RuntimeError(f"Failed to create directory: {save_path}")
                
            # 测试目录是否可写
            test_file = os.path.join(save_path, '.write_test')
            try:
                with open(test_file, 'w') as f:
                    f.write('test')
                os.remove(test_file)
            except Exception as e:
                raise RuntimeError(f"Cannot write to vector store directory: {str(e)}")
                
        except Exception as e:
            raise RuntimeError(f"Cannot create vector store directory: {str(e)}")
        
        # 创建空的向量存储
        try:
            # 创建一个空的文档列表来初始化向量存储
            empty_docs = [Document(page_content="")]
            vectorstore = FAISS.from_documents(empty_docs, self.embeddings)
            
            # 确保目录存在
            os.makedirs(save_path, exist_ok=True)
            
            # 先尝试创建一个临时文件以确保我们可以写入
            temp_file = os.path.join(save_path, "temp_index.faiss")
            with open(temp_file, 'w') as f:
                f.write("")
            os.remove(temp_file)
            
            # 保存向量存储
            vectorstore.save_local(save_path)
            print(f"Successfully initialized vector store at {save_path}")
        except Exception as e:
            raise RuntimeError(f"Failed to initialize vector store: {str(e)}")


    def load_vectorstore(self, load_path: str, trust_source: bool = False) -> FAISS:
        """
        Load a FAISS vector store from disk
        
        Args:
            load_path: Path to the vector store
            trust_source: If True, allows deserialization of the vector store.
                        WARNING: Only set to True if you trust the source.
        
        Returns:
            FAISS vector store instance
            
        Raises:
            SecurityError: If trust_source is False
            RuntimeError: If loading fails
        """
        # 转换为绝对路径
        load_path = os.path.abspath(load_path)
        
        if not os.path.exists(load_path):
            raise FileNotFoundError(f"Vector store not found at {load_path}")

        if not trust_source:
            warnings.warn(
                "Loading vector stores requires deserializing pickle files, which can be unsafe. "
                "If you trust the source of this vector store (e.g., you created it), "
                "set trust_source=True. Never set trust_source=True with files from untrusted sources.",
                UserWarning
            )
            raise SecurityError("Refusing to load vector store without explicit trust_source=True")

        try:
            required_files = ["index.faiss", "index.pkl"]
            missing_files = [name for name in required_files if not os.path.exists(os.path.join(load_path, name))]
            if missing_files:
                raise FileNotFoundError(
                    f"Vector store files missing in {load_path}: {', '.join(missing_files)}"
                )

            # FAISS on Windows can fail to read non-ASCII paths.
            # If the source path is not ASCII-safe, copy required files to an ASCII temp dir first.
            if not load_path.isascii():
                ascii_temp_dir = self._get_ascii_safe_temp_dir()
                try:
                    for file_name in required_files:
                        src = os.path.join(load_path, file_name)
                        dst = os.path.join(ascii_temp_dir, file_name)
                        shutil.copy2(src, dst)
                    return FAISS.load_local(
                        ascii_temp_dir,
                        self.embeddings,
                        allow_dangerous_deserialization=True
                    )
                finally:
                    shutil.rmtree(ascii_temp_dir, ignore_errors=True)

            return FAISS.load_local(load_path, self.embeddings, allow_dangerous_deserialization=True)
        except RuntimeError as e:
            raise RuntimeError(f"Failed to load vector store from {load_path}. Error: {str(e)}")

class SecurityError(Exception):
    """Raised when attempting unsafe operations without explicit permission"""
    pass