ASF-RAG-backend/RAG_M/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv
langchain
langchain-core
langchain-ollama
faiss-cpu
sentence-transformers
unstructured
python-multipart>=0.0.5
pypdf
pandas
openpyxl
langchain-huggingface
langchain-community
docx2txt
python-docx
langchain_ollama
google-api-python-client
google-auth-httplib2
google-auth-oauthlib